We just refactor the code, so that it works with webpack
