+++ src/index.ejs
@@ -3,7 +3,6 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="css/styles.css">
   <link href="https://fonts.googleapis.com/css2?family=Raleway:ital,wght@0,200;0,300;0,400;1,300&display=swap" rel="stylesheet">
   <title>TIC TAC TOE</title>
 </head>
@@ -33,6 +32,5 @@
     <button class="button" id="startGame">Start Game</button>
   </main>
 
-  <script src="js/script.js"></script>
 </body>
 </html>
